requirements-dev.txt
-r requirements.txt
flake8==3.7.9
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-sugar==0.9.2
requests==2.23.0
